create common prebuild check
